requirements.txt
streamlit
nltk
PyPDF2
python-docx
pandas
